+++ markitup/widgets.py
@@ -3,11 +3,15 @@ from __future__ import unicode_literals
 import posixpath
 from django import forms
 from django.contrib.admin.widgets import AdminTextareaWidget
-from django.core.urlresolvers import NoReverseMatch, reverse
 from django.template.loader import render_to_string
+try:
+    from django.urls import NoReverseMatch, reverse
+except ImportError:
+    from django.core.urlresolvers import reverse, NoReverseMatch
 from django.utils.safestring import mark_safe
 from markitup import settings
 from markitup.util import absolute_url
+
 
 
 class MarkupInput(forms.Widget):
@@ -71,7 +75,10 @@ class MarkItUpWidget(MarkupTextarea):
     def render(self, name, value, attrs=None):
         html = super(MarkItUpWidget, self).render(name, value, attrs)
 
-        final_attrs = self.build_attrs(attrs)
+        # Passing base_attrs as a kwarg for compatibility with Django < 1.11
+        # (where it will be treated as an innocuous attr named base_attrs)
+        final_attrs = self.build_attrs(
+            base_attrs=self.attrs, extra_attrs=attrs)
 
         try:
             preview_url = reverse('markitup_preview')