+++ CHANGES.rst
@@ -4,7 +4,8 @@ CHANGES
 master (unreleased)
 -------------------
 
-* none yet
+* Add support for Django 1.11 and 2.0.
+* Drop compatibility with Django 1.10 and lower.
 
 3.0.0 (2016.09.04)
 ------------------